delete useless code
